Correct string size
I misunderstood the string concept previously. If I declare a string in
a function, sizeof will return the string lenght including the null
character. However, when I use string for a function argument, sizeof
will return the memory address size. It passed previously because the
size it allocated is enough. However, I changed this for more accurate
string size.